+++ src/Schema/AbstractSchema.php
@@ -141,18 +141,6 @@ abstract class AbstractSchema implements SchemaInterface
         };
     }
 
-    /** @deprecated Use {@see Quoter::getRawTableName()}. Will be removed in version 2.0.0. */
-    public function getRawTableName(string $name): string
-    {
-        if (str_contains($name, '{{')) {
-            $name = preg_replace('/{{(.*?)}}/', '\1', $name);
-
-            return str_replace('%', $this->db->getTablePrefix(), $name);
-        }
-
-        return $name;
-    }
-
     /**
      * @throws InvalidArgumentException
      * @throws NotSupportedException
@@ -331,8 +319,7 @@ abstract class AbstractSchema implements SchemaInterface
      */
     public function refreshTableSchema(string $name): void
     {
-        /** @psalm-suppress DeprecatedMethod */
-        $rawName = $this->getRawTableName($name);
+        $rawName = $this->db->getQuoter()->getRawTableName($name);
 
         unset($this->tableMetadata[$rawName]);
 
@@ -462,8 +449,7 @@ abstract class AbstractSchema implements SchemaInterface
      */
     protected function getTableMetadata(string $name, string $type, bool $refresh = false): mixed
     {
-        /** @psalm-suppress DeprecatedMethod */
-        $rawName = $this->getRawTableName($name);
+        $rawName = $this->db->getQuoter()->getRawTableName($name);
 
         if (!isset($this->tableMetadata[$rawName])) {
             $this->loadTableMetadataFromCache($rawName);
@@ -543,11 +529,8 @@ abstract class AbstractSchema implements SchemaInterface
      */
     protected function setTableMetadata(string $name, string $type, mixed $data): void
     {
-        /**
-         * @psalm-suppress MixedArrayAssignment
-         * @psalm-suppress DeprecatedMethod
-         */
-        $this->tableMetadata[$this->getRawTableName($name)][$type] = $data;
+        /** @psalm-suppress MixedArrayAssignment */
+        $this->tableMetadata[$this->db->getQuoter()->getRawTableName($name)][$type] = $data;
     }
 
     /**